requirements.txt
Django>=6.0.2
google-genai>=0.2.0
python-dotenv>=1.0.0

